Under he --new-profile-management flag, migrate the desktop profile avatars (hamburger, ninja) to the CrOS ones (penguin, dog, blurry circles).
Regarding the avatar resource files: I will have a followup CL to move the
resource files to /common, once this lands. For now we'd like to monitor
how including these files affects the size of the binary, and if it
increases too much, have an easier revert.

BUG=312352

Review URL: https://codereview.chromium.org/60263002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234195 4ff67af0-8c30-449e-8e8b-ad334ec8d88c