[win8] Keep new windows for each packaged app in Ash or Desktop, and switch modes when we can.
Currently multi-window packaged apps can spread their windows over both
desktop types. Also, for "singleton" apps that already have a window in
the non-active desktop, attempting to launch it again appears to do
nothing, since it activates in a hidden desktop.

With this change, when a packaged app already has a window,
NativeAppWindowViews will ensure that subsequent windows are opened in
the same desktop.

When launching an app from Ash that already has a window open on the
Desktop, we instead ask the metro viewer process to launch the app as if
using a desktop shortcut for the app. This allows Chrome to
automatically switch to desktop mode (using the only way Windows 8 will
let a process do this; via ShellExecuteEx).

When launching an app from Desktop mode that has a window in Ash, or
when activating or showing a window in Ash when Ash is not the active
desktop, Chrome will automatically switch into metro mode.

BUG=288251

Review URL: https://codereview.chromium.org/33123005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234470 4ff67af0-8c30-449e-8e8b-ad334ec8d88c