History: Fix and re-enable the singleDeletion browser test.
The problem was that the test was waiting for any element to be deleted
from the DOM, but deleting a single history entry would result in several
DOM elements being deleted. The fix is to add a callback that is only
called when everything has been deleted.

BUG=242293
R=bauerb@chromium.org

Review URL: https://codereview.chromium.org/64123003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234477 4ff67af0-8c30-449e-8e8b-ad334ec8d88c