Policy providers all get a SchemaRegistry to work with.
The SchemaRegistry contains the components that the policy provider should load
policy for, as well as the expected policies and their types for each component.

Also killed the PolicyDomainDescriptor, which was a similar way of doing this
but had the problem of overlapping descriptors on shared policy providers.

BUG=270667
R=bartfab@chromium.org, jochen@chromium.org, kalman@chromium.org

Review URL: https://codereview.chromium.org/56623005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234514 4ff67af0-8c30-449e-8e8b-ad334ec8d88c