BMM: Fix issue with disabled context menu items for search "folder".
Revert of revert r233314

Update command state on "load" of a folder.

We used to only update the command when selection changes. However, when
a folder is loaded the selection does not change, it goes from [] to []
so no event is fired. However, we do have callbacks after the folder is
populated. Use that to update the state of all the commands.

BUG=313592
TBR=kbr@chromium.org

Review URL: https://codereview.chromium.org/65303007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c