Remove 4 static initializers on linux.
clang doesn't need static initializers for TimeDeltas, but gcc does. This
accounts for 2 initializers.  Also move two constants to the header which are
used in the initialization of other constants ("const int kB = kA + 10"). This
accounts for the other 2 (one of them being introduced in the aura switch.)

BUG=317171,94925
R=erg@chromium.org

Review URL: https://codereview.chromium.org/61643020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234901 4ff67af0-8c30-449e-8e8b-ad334ec8d88c