Linux users can now turn on --apps-use-native-frame.
Unfortunately, it does not currently produce an actual native frame (it
results in a blue Aura window). But it does have a different effect to
the default frame (which is the same as the default frame on Windows) so
it is worth having the option to play with.

BUG=225916

Review URL: https://codereview.chromium.org/70973005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@235025 4ff67af0-8c30-449e-8e8b-ad334ec8d88c