Fix windowed NPAPI plugins appearing under dialogs when DWM compositing is disabled and the tab switches back and forth.
I fixed the one-tab case in r234823, but while testing I found that if I switched to another tab and and came back, then the fix stopped working. The bug was that when the root window became the parent, the next time the parent window changed we stopped observing the parent window, but never observed the root window again.

BUG=299224
R=ben@chromium.org

Review URL: https://codereview.chromium.org/73523002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@235305 4ff67af0-8c30-449e-8e8b-ad334ec8d88c