Add remove account UI to the new avatar bubble.
Side note: this only allows the deletion of secondary accounts.

BUG=311111
TEST=With the --new-profile-management and --enable-inline-signin
flags, start chrome, and sign in a profile. Through the avatar bubble
add a secondary account. Once it's added to the list, you should be able
to see a little x next to it, and pressing it will delete it from the list.

Review URL: https://codereview.chromium.org/60733026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@235476 4ff67af0-8c30-449e-8e8b-ad334ec8d88c