Refactor some ProfileSyncServiceTests
Create a new test framework class named ProfileSyncServiceSimpleTest and
use it to replace several ProfileSyncServiceTests.  This newer framework
takes advantage of MockSyncBackendHost and allows us to use a real
ProfileSyncService in these tests.

The resulting tests are better at testing the ProfileSyncService, but
they no longer test any part of the SyncBackendHost code.   This is a
desirable trade off because these are unit tests.

BUG=312994

Review URL: https://codereview.chromium.org/69583002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@235661 4ff67af0-8c30-449e-8e8b-ad334ec8d88c