ComponentCloudPolicyService tracks the signin state.
The ComponentCloudPolicyService is used to fetch policy for Chrome components
such as extensions. This change makes it follow the sign-in state by observing
the cloud policy store and a cloud policy core, which determine that state
from the cloud policy system perspective.

The component policy cache will be cleared on signout.

The ComponentCloudPolicyService now tracks a CloudPolicyCore to determine
whether a CloudPolicyClient is available or not; it is possible to signin
and signout multiple times.

BUG=108992

Review URL: https://codereview.chromium.org/72793003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@235747 4ff67af0-8c30-449e-8e8b-ad334ec8d88c