Fix reporting of Extensions.KnownDisabledReDisabled to also report 0 when there are no known_disabled extensions.
As it is now we don't report neither Extensions.KnownDisabledReDisabled:0, nor Extensions.KnownDisabledInitialized, when there are no known_disabled and no currently disabled_extensions.

This CL makes it so that we always report either and will result in an increase of the number of 0 reports for Extensions.KnownDisabledReDisabled.

Review URL: https://codereview.chromium.org/76533002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236109 4ff67af0-8c30-449e-8e8b-ad334ec8d88c