net-internals: hide 'original proxy' if equal to effective proxy
When both are equal the display is somewhat visually noisy, so hide the original
proxy div if the strings are equal and show it again when they begin to differ.
You can test this by pressing 'reapply' on the proxy page, which will
temporarily cause the original and effective proxy to differ.

BUG=235998
TEST=adhoc

Review URL: https://codereview.chromium.org/61953004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236137 4ff67af0-8c30-449e-8e8b-ad334ec8d88c