Cancel all in-progress downloads when the last browser window is closed.
When quitting using wrench quit or when the last browser window is
closed on Windows and Linux, a dialog warns that all downloads will be
cancelled. Previously, if a packaged app was open at the time, the
downloads continued. With this change, on those platforms, when the last
browser window is closed, all in-progress downloads are cancelled.

BUG=320592

Review URL: https://codereview.chromium.org/73583005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236148 4ff67af0-8c30-449e-8e8b-ad334ec8d88c