Fix Files app popping up after opening the lid with a sd/usb device in.
This patch fixes a race between the resuming and mounting events causing failures in detecting remounting.
Since we may get the resuming event after mounting, we have to flip the remounting flag during suspending.
If we do so, then once the lid is open, we can detect that we are resuming immediately.

TEST=Tested manually on Butterfly.
BUG=311901

Review URL: https://codereview.chromium.org/77463002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236208 4ff67af0-8c30-449e-8e8b-ad334ec8d88c