Fix a crash when chrome.downloads.removeFile called on invalid download.
If the DownloadsRemoveFileFunction constructor doesn't explicitly set item_=NULL, then it's set to non-NULL garbage. If the item id passed to removeFile is invalid, then item_ is not set to a live DownloadItem*, so that non-NULL garbage is dereferenced in the destructor item_->RemoveObserver(this).

Re-enable tests on mac and linux since flakiness was only reported on windows.

BUG=319984,306144

Review URL: https://codereview.chromium.org/76393004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236294 4ff67af0-8c30-449e-8e8b-ad334ec8d88c