Move ProxyPolicy fixup logic out of PolicyProvider and inject it instead.
The logic was previously applied at the UpdatePolicy level of every
ConfigurationPolicyProvider, but this needs to move to facilitiate the
refactoring of chrome/browser/policy into a layered component. Instead
a fixup function can be injected into a PolicyService at construction
time and is applied before policies from the various PolicyProviders
are merged.

There are three tests for this:
- PolicyTransformationsTest.FixDeprecatedPolicies tests that the proxy
  policy fixup function works properly.
- PolicyServiceTest.PolicyPreprocessing tests that the preprocessor
  function is applied during policy merge.
- PolicyTest.PolicyPreprocessing tests that the proxy fixup function
  is added BrowserPolicyConnector and ProfilePolicyConnector.

TEST=unit test, integration test, browser test
BUG=271392

Review URL: https://codereview.chromium.org/59243005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236511 4ff67af0-8c30-449e-8e8b-ad334ec8d88c