Ensure the ExtensionSystem is ready before initializing an ExtensionAppModelBuilder
Some code paths currently result in an AppListSyncableService being
created for a profile before the profile is fully initialized. In this
case, the ExtensionSystem can be used, but the Extension*Service* is not
yet created, and has not finished loading all the extensions. Attempting
to populate the app launcher model at this point results in no apps
being found.

This CL adds a check in AppListSyncableService to see if the
ExtensionSystem is ready. If not, wait for it to become ready and
populate apps then.

BUG=321514
TEST=(On mac)Have two signed-in profiles; Have app launcher showing
Profile "A", Close all browser windows except for a browser showing
profile "B", Quit Chrome, Show app launcher [Chrome starts], Switch app
launcher to profile "B". App launcher should show apps, and not be
blank.

Review URL: https://codereview.chromium.org/77773002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236513 4ff67af0-8c30-449e-8e8b-ad334ec8d88c