AppListPositioner no longer needs a shelf_rect.
The new, more general, interface allows the client to set up the work
area as desired, and then positions the app list within the work area.

This doesn't change much on Windows, but on Mac and Linux it will mean
we don't need to compute the shelf rectangle from the work area just to
give to AppListPositioner.

BUG=317944

Review URL: https://codereview.chromium.org/81243004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236741 4ff67af0-8c30-449e-8e8b-ad334ec8d88c