Add an accessibility alert for incorrect use of some accelerators.
Some accelerators don't make sense if no windows exist on the desktop.
ChromeVox users can end up in this state and not know it.
If a window-needing accelerator is issued and there is no window we
fire off an accessibiliy event to let the user know there is no
window.

This CL:
Extends Ash's AcceleratorController to intercept these accelerators.
Adds the concept of triggering an alert to AccessibilityDelegate.
Adds handling in Chrome's AccessibilityDelegate to generate an alert.

BUG=217571
TEST=All automated tests pass. Activate ChromeVox and hit ctrl-m with
no window open on Chrome OS. You should get an alert telling you to
open a window.

Review URL: https://codereview.chromium.org/77453013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236812 4ff67af0-8c30-449e-8e8b-ad334ec8d88c