Disable the Manage SSL certificates button under Win Ash.
This is necessary since we don't have a native Ash replacement for the ssl
certificate dialog, and we need to be able to ship Ash with M-32.  We plan to
integrate support with the Win8 system charm at some point in the future, but
it's a significant chunk of work and we need to do something so that Chrome
doesn't become unusable if the user clicks this button.

BUG=255329

Review URL: https://codereview.chromium.org/79223003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@236893 4ff67af0-8c30-449e-8e8b-ad334ec8d88c