Files.app: Fix summarizing of progress items.
In the closed view of the progress center, a single item should be shown
directly and multiple items should be shown as a summirized item.
Originally the summarizing method had bugs such that:

1. 1 progressing item and 1 canceled item are summarized. But we should show 1
progressing item directly because a user does not concern canceled items.

2. 2 error items are summirized as an item having progressing state. But it
should be summirized as an item having error state.

This patch turns the method to the currect behavior.

Along the way, this patch also fixes a behavior of toggle button
to prevent the button acting as a cancel button.

BUG=321483
TEST=manually

Review URL: https://codereview.chromium.org/80133003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@237036 4ff67af0-8c30-449e-8e8b-ad334ec8d88c