Implement new invalidations ack tracking system
Enable the new invalidation ack tracking API.  This system allows the
invalidations system to remember invalidations that have not been acted upon
yet across browser restarts.

New features include:
- Supports saving multiple invalidation payloads per ObjectID.
- Supports tracking lost or dropped invalidations.
- Remembers "unknown version" invalidations across restarts.
- Can send multiple payloads in a single callback.
- Does not periodically ping unacked invalidations.

A series of commits prior to this one have put much of the framework in place
already.  This CL does not change the types and signatures of client-facing
APIs very much, but it does enable perviously disbaled behavior.  This CL
includes updates to the client code to allow them to support the new semantics
with (hopefully) few changes to externally visible behavior.

Most of the big changes are in the classes and tests associated with the
internals of Invalidations.  Large parts of the SyncInvalidationListener and
InvalidatorStorage have been rewritten.

BUG=233437

Review URL: https://codereview.chromium.org/56113003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@237421 4ff67af0-8c30-449e-8e8b-ad334ec8d88c