Reland review 34393006: Refactor MediaStreamManager to not output real device id. https://src.chromium.org/viewvc/chrome?revision=232766&view=revision
Refactor MediaStreamManager to never output physical device ids. It now output sourceId in the form of a HMAC.
DeviceMessageFilter now don't need to create a source id.
This also fix a bug that made the source ids useless unless you have called gum once.
Note that the sourceIds are still just using the security origin as salt. Next step is to add a random number that is stored in the profile.

BUG= 269139,313192

Review URL: https://codereview.chromium.org/88283002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@237825 4ff67af0-8c30-449e-8e8b-ad334ec8d88c