Fix dragging supported files to tab strip.
Use the inferred MIME type of download items, rather than the server value,
when deciding if the tab strip can allow a drop. Get the MIME type
asynchronously so we don't crash (IO must be allowed in the call to
GetMimeTypeFromFile).

Also check the profile's plugins when determining whether a MIME type is
supported.

Regardless of whether a file is supported, the dragging operation should
still be allowed so that spring-loaded tabs work with file drags.

BUG=274288
R=asanka@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/68133020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@238253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c