Stop setting invalid value to the |tab_id| field of fileBrowserHandler API.
Since the very long long ago when we have moved file manager out of tabs
to a separate app window, it didn't have any tab_id associated,
and the existing code always assigned an invalid value (-1).

This CL removes the useless code. Since the field is defined to
be optional, it doesn't change the interface of the API.

BUG=312519

Review URL: https://codereview.chromium.org/101323002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@238376 4ff67af0-8c30-449e-8e8b-ad334ec8d88c