Ensure that reposting mouse events on Windows posts the message natively via the PostMessage API.
The current code posts the events via the Aura code path which does not work on Windows with the glass theme for the system
menu buttons (Min/Max/Restore). These are part of the non client area of the Window and rely on the WM_NCLBUTTONDOWN message
to be defproced which eventually posts a WM_SYSCOMMAND message for the corresponding operation.

Fix is to check if the NativeWindow doing a repost is in the Metro environment for Windows. If yes we continue the current logic
of reposting the events via Aura. If not we post the message natively. The RepostLocatedEvent function internally calls
the RepostLocatedEventWin function to post the function natively. The functionality to check if a window is in metro
has been added to the ViewsDelegate interface and is implemented by Chrome.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=323105

BUG=323105
R=sky@chromium.org

Review URL: https://codereview.chromium.org/100303005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@238462 4ff67af0-8c30-449e-8e8b-ad334ec8d88c