Fix issues with token refresh in AccountReconcilor
The CL was change a bit as suggested by the reviewers.

This CL now introduces a worklist model in GoogleAutoLoginHelper. Instead of firing all the refresh token events in a row, it queues up all the login requests and insert GAIA cookies one-by-one. This would let accounts to be appended instead of overwriting one another.

This also introduces GoogleAutoLoginHelperTest for some unit testing.

BUG=305249

Review URL: https://codereview.chromium.org/63253003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@238547 4ff67af0-8c30-449e-8e8b-ad334ec8d88c