Add tests for UserImageManager
This CL adds tests that verify the various methods for choosing the user
image set and persist the image correctly. One of the tests also verifies
that when the profile image download running in the background completes,
the downloaded image is not used if the user has chosen a different image
type in the meantime.

BUG=None
TEST=Expanded browser test coverage

Review URL: https://codereview.chromium.org/96473005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@238845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c