Pass "shared" metadata of Drive entries in fileBrowserPrivate API.
This information is necessary for distinguishing shared folders in Files.app UI.
The similarly named "shared_with_me" field has a different meaning that is also
necessary, hence we need to pass both. (The term and the semantics is derived
from Drive API; we're keeping it as-is for consistency. For the detail of the
difference, see comments in c/b/cros/drive/drive.proto.)

BUG=264812

Review URL: https://codereview.chromium.org/103883002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@238856 4ff67af0-8c30-449e-8e8b-ad334ec8d88c