Add UMA stats for discarded tabs on a per logged in user base
Each discarded tab will be "stored" in the bucket for the number of users logged in at the time of the occurrence. 
To get to a useful number, the value needs to be normalized by dividing the absolute value per bucket by the corresponding UsersPerSession bucket (in percent).

BUG=321739
TEST=visual

Review URL: https://codereview.chromium.org/104713003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@239023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c