Ensure that the Chrome downloads folder option in Chrome ASH on Windows switches to desktop mode while opening the downloads folder.
The current behavior is that the downloads folder is opened on the desktop while the user continues to stay in ASH mode
leading to confusion.

To fix this we reuse the ActivateDesktopHelper function in application_lifetime_win.cc. This function takes a boolean parameter
now which indicates if we should be exiting ash after activating the desktop. This flag is true when we launch desktop chrome from
ASH and is false when invoked to open the downloads folder via the functions in platform_util_win.cc. The MetroViewerHostMsg_ActivateDesktop
IPC now has a flag which indicates if we should exit ash after activating desktop. If this flag is false we terminate the
delegate_execute process which was launched to switch to desktop.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=324252

BUG=324252
R=cpu@chromium.org, jschuh@chromium.org, sky@chromium.org, cpu

Review URL: https://codereview.chromium.org/102603006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@239041 4ff67af0-8c30-449e-8e8b-ad334ec8d88c