Reinstate alwaysOnTopWindows permission
Reinstate the "alwaysOnTopWindows" permission. This feature is 
available only in the Dev channel, but is whitelisted for two first-party
apps in Stable.

Apps that wish to enable the alwaysOnTop property for app windows must
now declare the "alwaysOnTopWindows" permission in their manifest.
This permission does not generate a warning in install prompts.

BUG=326361
TESTS=browser_tests (PlatformAppBrowserTest.WindowsApiAlwaysOnTop*)
Manual test: Create an app which does not specify this permission and
attempts to create a new window with the option enabled or calls
setAlwaysOnTop(true) on an existing window. The window should not be
always on top. The option should work for apps which do declare the
permission.

Review URL: https://codereview.chromium.org/61073004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@239180 4ff67af0-8c30-449e-8e8b-ad334ec8d88c