Add a PDF component extension
This adds a PDF component extension which is used for rendering PDFs with an
out of process PDF plugin. It adds the extension as a MIME handler for pdf file
types. The extension is currently only added if the --out-of-process-pdf switch
is passed to Chrome, AND the PDF plugin is available.

BUG=303491

Review URL: https://codereview.chromium.org/96233002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@239451 4ff67af0-8c30-449e-8e8b-ad334ec8d88c