Settings: Can't activate URL fields in 'Default search settings' section.
This is a regression by r239172.

Bug detail:
These URL fields are input[disabled], which are not focusable. After r239172, a
static version of a URL field gets focus by mousedown, and we try to move focus
to the URL field later. But the URL field is not focusable and we loses
focus. This behavior worked before r239172 because the owner <list> got focus by
mousedown.

Fix:
We focus on the owner <list> explicity in such case. It's same as r239171 or
earlier.
Also, this CL fixes a disabled element issue in containsFocusableElement().
This doesn't affect any behavior, and it just improves correctness.

BUG=326863

Review URL: https://codereview.chromium.org/109943002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@239726 4ff67af0-8c30-449e-8e8b-ad334ec8d88c