[OriginChip] If the field trial is enabled, omit the URL from the omnibox.
This is done by reusing the mechanisms established by the query_extraction field trial. The 'Show URL' action from the right-click menu is shown when either query_extraction or origin_chip is enabled.  If query terms are not shown, then an empty string is returned as the text from the toolbar model.

BUG=315944

Review URL: https://codereview.chromium.org/93233002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@239845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c