Working around the page load problem on Linux.
After experimenting on the bots, this admittedly silly workaround seems
to work without any negative side effects. This patch makes the main browser test and
the video quality test use the same helper to load pages, and then makes that helper load
the page twice on Linux. I tried that on the bot and it manages to load the js
correctly 100% of the time if I do that.

BUG=281268

Review URL: https://codereview.chromium.org/99333010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@240170 4ff67af0-8c30-449e-8e8b-ad334ec8d88c