Fix hosted app and extension windows using low resolution taskbar icons.
Hosted app and extension windows currently use the favicon as the
taskbar icon. This means the taskbar icon is scaled from 16x16 to 32x32
which results in an ugly image. This CL makes app windows use the
icon of the extension that URL is associated with.

The retrieved icon size has been changed from 24x24 to 32x32 in
order to avoid bad upscaling in the taskbar and higher resolution icons
will be downscaled if no 32x32 icon is provided.

BUG=94301

Review URL: https://codereview.chromium.org/113293004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@240644 4ff67af0-8c30-449e-8e8b-ad334ec8d88c