Reland of https://codereview.chromium.org/114063003.
In addition to the previous patch, removing a test that was failing.
The test was testing behavior that is about to go obsolete anyway. The test
essentially tests that the gUM goes through even though it can't acquire both
audio and video. This is behavior that will change shortly, and the particular
test violates the new checks I added which are intended to spot when the
webcam is broken, so I thought it made sense to just remove the test.

Make WebRTC browser tests fail cleaner when the host fails to provide a camera.

BUG=327673
NOTRY=True

Review URL: https://codereview.chromium.org/109303005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@240910 4ff67af0-8c30-449e-8e8b-ad334ec8d88c