[sync] Simplify EnableEncryption and WaitForEncryption
ProfileSyncServiceHarness provides sync integration tests with a couple
of methods to enable / wait for encryption. These methods inspect internal
state in ProfileSyncService to determine when encryption is complete.

This CL removes the reliance on sync session snapshots in
WaitForEncryption. In addition, it also removes the condition that
IsFullySynced be true before WaitForEncryption returns true. This is
done under the premise that tests are responsible for waiting for sync
to complete after encryption is enabled, by calling AwaitQuiescence or a
related Await* method. An audit of all the integration tests that enable
encryption reveals that this is already true.

BUG=95619,323380
TEST=sync_integration_tests:*Encryption*

Review URL: https://codereview.chromium.org/115523002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@240985 4ff67af0-8c30-449e-8e8b-ad334ec8d88c