[rAC, OSX] Force better initial size for sign-in (2nd try)
This is a re-land of http://crrev.com/103853004 (r240566), which was
reverted in http://crrev.com/113913008 (r240603)

The revert was due to a test failure with the root cause that
AutofillDialogBrowserTest shows the dialog, but never runs the
NSRunLoop - which means the visual display is not fully updated. And
showing the sign-in dialog requires the main dialog's size to be non-
zero.

Fixes added:
* AutofillSignInDelegate DCHECKs for invalid sizes (better callstacks)
* AutofillDialogBrowserTest cycles run loops whenever dialog is shown

---- original description ----
Open the sign-in dialog with an initial size equal to the current
dialog. This avoids one unnecessary resize until the content is loaded
and makes the visual experience less jarring.
---- end original description ----

TBR=jochen@chromium.org
R=isherman@chromium.org
BUG=325518

Review URL: https://codereview.chromium.org/97993006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241009 4ff67af0-8c30-449e-8e8b-ad334ec8d88c