sync: Use invalidation local acknowledgements
As of r237421, the invalidations system now supports a new and improved
form of local acknowledgements.  Any invalidations not explicitly
dropped or acknowledged will be redelivered after a restart.

This CL makes use of the new functionality in sync.  It uses the
DataTypeTrackers to track the status of pending invalidations and
acknowledge or drop invalidations as appropriate.

Also notable in this CL is that the 'server_dropped_hints' flag has now
been uncommented in the sync protocol.  We've known for a long time that
we wanted to support such a flag, but haven't been able to implement it
correctly until now.

BUG=78462,233437

Review URL: https://codereview.chromium.org/102353005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241170 4ff67af0-8c30-449e-8e8b-ad334ec8d88c