Allow built-in bookmarks extension to use page action for bookmark star
Allow Chrome's built-in bookmarks extension to replace the bookmark star
with its page action. This provides more flexibility for experimenting
with the appearance of the bookmark button in the omnibox. We permit
this behavior based on the bookmark manager private permission since we
may experiment with multiple test extensions and all require this
permission.

This CL is composed of two separate changes:
 - allow use of the hide bookmark star extension API
 - allow moving the page action to the rightmost position in the omnibox

BUG=311282
R=shess@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/114333003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241171 4ff67af0-8c30-449e-8e8b-ad334ec8d88c