Linux Aura: Browser window caption buttons now extend to top of screen.
This now obeys Fitts' Law (the top row of pixels are clickable for
minimize, maximize and close buttons, and the top-right corner is
clickable for the close button).

Visually, the buttons' position is unchanged; just their click region
has been extended to the top of the screen.

BUG=318920
TEST=On non-Unity window manager with no bar along the top (e.g., KDE or
Cinnamon), maximize Chrome. Minimize and maximize should be clickable
with the cursor along the top edge of the screen; close should be
clickable with the cursor in the corner pixel.

Review URL: https://codereview.chromium.org/108563013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241191 4ff67af0-8c30-449e-8e8b-ad334ec8d88c