Dynamic labels for the undo/redo context menus in bookmark bar.
The labels for the Undo and Redo context menu items change based on
the next action. For example, when a bookmark is deleted the menu item
will show as "Undo delete" and "Redo delete".

This is visible with the switch enable-bookmark-undo and found in the
bookmark bar context menu.

BUG=126092

Review URL: https://codereview.chromium.org/115233005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241309 4ff67af0-8c30-449e-8e8b-ad334ec8d88c