Android: allow to set the initial visibility of WebContents being created.
This patch exposes the |initially_hidden| creation parameter of WebContents in
Java ContentViewUtil, allowing the Android embedder to correctly set the
visibility of WebContents created for tabs opened to load in background.

Also, the |are_layers_attached| flag of RenderWidgetHostViewAndroid is now set
in constructor to match visibility tracked in RenderWidgetHost.

This allows to fix the Page Visibility API for tabs opened to load in
background.

BUG=327125

Review URL: https://codereview.chromium.org/110873006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241399 4ff67af0-8c30-449e-8e8b-ad334ec8d88c