Chrome side of making SharedWorkerRepositoryClient be per RenderFrame instead of RenderView.
This also fixes the TODOs from r241306 in tab_specific_content_settings.cc

The RenderView ID is still being sent to the browser because it's used by ResourceRequestInfo::GetAssociatedRenderView. Fixing that will be a large change by itself, so I'll do that in a followup.

One minor behavior change is that the per-RenderView limit on how many shared workers can be started is now per-RenderFrame. Given that shared workers will be moved to the renderer process, I'm not too concerned about that.

BUG=304341
R=nasko@chromium.org

Review URL: https://codereview.chromium.org/109213006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241514 4ff67af0-8c30-449e-8e8b-ad334ec8d88c