Fix pointer lock tests for Chrome Apps.
The focus was not guaranteed by using the ui_test_uitls call.
It was made obvious by cl 108003007 (the cl made the document no longer
focused if the window is not focused which was the case here and as a
consequence, the test was failing).

BUG=223447

Review URL: https://codereview.chromium.org/113343005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241547 4ff67af0-8c30-449e-8e8b-ad334ec8d88c