Start device policy core if policy is present.
Previously, we'd only start if the policy indicated that the client is
actively managed. That's wrong because a device in state UNMANAGED
needs to be able to fetch policy to get out of that state again. Also
add a new unit test to cover this scenario.

BUG=chromium:329203

Review URL: https://codereview.chromium.org/98143009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241556 4ff67af0-8c30-449e-8e8b-ad334ec8d88c