Change initial download path preference on Chrome OS to multi-profile aware source.
The current default is "$HOME/Downloads", which is "/home/chronos/user/Downloads".
This CL is preparation for the switch to "profile->GetPath()/Downloads".
(I'll trigger the actual switch by changing the implementation of
fm::util::GetDownloadFolderForProfile() once all the migration patches landed.)

The switch makes difference in the case of multi-profile sessions.
Since "/home/chronos/user" is always a hard link to "primary_profile->GetPath()",
if "/home/chronos/user/Downloads" is stored in the preference, when a profile
is loaded as a secondary profile, it'll try to save to primary_profile/Downloads,
not the secondary one's own Downloads folder.

Note: I'm not trying to purge all DownloadPref::GetDefaultDownloadDirectory();
the uses of the function as the last-resort destination is still valid because
"/home/chronos/user/Downloads" is still available. What I want here is to *not*
to store the /user/ path to the preference.

BUG=309556

Review URL: https://codereview.chromium.org/54383002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c