Run all flag-checks for new-profile-managment et al through common functions for future Finch integration.
In order to do tests and a safe, controlled roll-out of Chrome's new profile management, we're going to enable it using a Finch experiment.  The current tests of the existing command-line flags all need to be routed through common functions in order to override these settings if a Finch experiment is set.

Affected flags are:
  --new-profile-management
  --enable-inline-signin
  --google-profile-info

Note that there existed a function to test for NewProfileManagement but it was not in a good place for general testing and definitely not for other flags so it has been extracted as well.

BUG=324046

Review URL: https://codereview.chromium.org/108803007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c