[Downloads] Use GetPluginInfo to determine if a MIME type is handled.
When determining whether a download file would be handled safely by the
browser, the downloads looks for a plug-in that handles the file's MIME
type. This determination was done using GetPluginInfoArray. The
resulting list of plugins had to be manually checked to make sure
disabled plug-ins weren't considered. This CL changes the behavior so
that the code uses GetPluginInfo instead. Doing so brings the code
closer to how BufferedResourceHandler performs the same check.

BUG=none

Review URL: https://codereview.chromium.org/90693002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241732 4ff67af0-8c30-449e-8e8b-ad334ec8d88c