Reinstate linkable ephemeral apps experiment
Reverts https://codereview.chromium.org/100553002
Reason for revert: The reverted patch was requested by the security
team, but there was a misunderstanding about whether this was a finch
experiment or a feature behind command line flags.

There is one difference with the original patch:
This patch restricts the ability to launch ephemeral apps to
https://www.google.com (previously allowed both http and https).

The "enable-linkable-ephemeral-apps" command line flag still must
be enabled.

BUG=312460
TBR=sky@chromium.org, benwells@chromium.org

Review URL: https://codereview.chromium.org/100133005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241807 4ff67af0-8c30-449e-8e8b-ad334ec8d88c