Extension in GCMProfileServiceRegisterTest.RegisterFromStateStore needs to get "gcm" permission
The extension used in the test uses an API not launched to stable (gcm).

This apparently caused a timeout on official builders, because official builds default to stable channel even if the code is ToT.

This CL uses ScopedCurrentChannel to set the channel to CHANNEL_UNKNOWN temporarily for loading the test extension. This results in getting the "gcm" permission.

This CL also enables the disabled test.

A considered alternative fix was changing the extension manifest's location to COMPONENT. See a similar change https://codereview.chromium.org/113993002/ for further discussion of the problem. The result is not to do it, because it has further reaching consequences and breaks other tests. Also adding the test extension into permissions features whitelist was abandoned in favour of the current solution.

BUG=326321

Review URL: https://codereview.chromium.org/99323011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c