mac: Fix a newly introduced history swiping bug.
Attempting to 2-finger history swipe twice in quick succession would
fail to register the second swipe ~50% of the time. The NSEvent callback with
phase=NSEventPhaseBegan non-deterministically occurs before the event has been
registered as a gesture. This was causing the history_swiper to process the
event as part of the previous gesture.

BUG=138175
TEST=In a single tab, navigate to 5 different sites. Use a trackpad to 2-finger
swipe repeatedly in the same direction. Each swipe should register as a history
swipe, and cause the page to navigate in the appropriate direction. There
should be no jankiness, and the touches should not fail to register.

Review URL: https://codereview.chromium.org/117733002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@241918 4ff67af0-8c30-449e-8e8b-ad334ec8d88c