Reset is_completed_ before attempting a new crx component download.
When downloading a url, the state of the BackgroundDownloader must be
reset.

A better solution later on would be implementing a BitsFetcher that can
only do one download, similar to UrlFetcher. This avoids the situation
when dirty state post-download is carried over to a subsequent
download.

BUG=304354

Review URL: https://codereview.chromium.org/102993014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@242108 4ff67af0-8c30-449e-8e8b-ad334ec8d88c