Getting rid of temporary IsGetDefaultProfileAllowed instrumentation
Getting closer to get rid of GetDefaultProfile. This is now removing the temporary instrumentation from stevenjb which was adding a check to see that GetDefaultProfile does not get called before a Profile was created. I discussed with dpolukhin and nkostylev and instead of renaming this to something like "WasProfileCreated" (or similar) the instrumentation should get removed (since it was temporary to begin with).

BUG=322682
TEST=covered by existing unit tests

Review URL: https://codereview.chromium.org/119593002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@242169 4ff67af0-8c30-449e-8e8b-ad334ec8d88c