Extract an ExtensionRegistry from ExtensionService
This new class holds the various sets of enabled/disabled/terminated/etc.
extensions. It lives in src/extensions/browser so it can be used by
app_shell.

This CL renames some members and methods to clarify which set of extensions is the enabled set and also renames some methods that are only used in test code.

BUG=none
TEST=unit_tests ExtensionRegistryTest and ExtensionServiceTest

Review URL: https://codereview.chromium.org/111873008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@242248 4ff67af0-8c30-449e-8e8b-ad334ec8d88c