Only close chrome://settings/clearBrowserData if it is visible
The chrome://settings/clearBrowserData overlay is closed when the clear
operation has finished. This CL ensures that we only close the overlay if it is
actually visible. If the user navigated to a different overlay, that overlay
should not be closed when the browsing data has been cleared.

BUG=330706
R=dbeam@chromium.org

Review URL: https://codereview.chromium.org/114863006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@242887 4ff67af0-8c30-449e-8e8b-ad334ec8d88c