mac: Fix a history swiper bug where the shield wouldn't disappear.
There were 3 separate problems.
1. Retain bug :(
2. historyOverlay was removing the view from the superview during dealloc,
which is atypical. If the view didn't exist, historyOverlay was actually
creating the view, and sending a callback to the delegate.
3. We were not specifying an NSAnimationContext grouping/duration, which
non-deterministically would use a 0-duration animation, which actually
prevented the animation and callback from triggering. As a result, the
shield would never disappear.

On a side note: unit_tests did not work when run with
--gtest-filter=HistoryOverlay* because we did not specify an NSAnimationContext
with a non-zero duration. 0-duration animation contexts cause animations to not
fire, which meant that the delegate callback never went through.

BUG=330312
TEST=manual

Review URL: https://codereview.chromium.org/107343008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@242911 4ff67af0-8c30-449e-8e8b-ad334ec8d88c