[Android] Fix the progress indicator when leaving native pages.
Force a progress update inside of LoadingStateChanged of the delegate.
When navigating between native pages, the load progress was not being
updated, which broke the case where you navigated to a slow page and
clicked a native page link while waiting.  The previous load progress
wouldn't be cleared and the 'x' in the URL bar would remain.

BUG=298396

Review URL: https://codereview.chromium.org/117243003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@242962 4ff67af0-8c30-449e-8e8b-ad334ec8d88c