Implement delete methods for the various CannedBrowsingDataHelpers.
This will allow individual cookies and other site data to be deleted via the
CookieTreeModel GUI. As a result of this change, updates to the various
storage systems will be pushed to the canned helpers after creation of the
CookieTreeModel.

Also removes a pimpl for BrowsingDataIndexedDBHelper since no other canned
helpers followed this pattern.

BUG=168996
TEST=No visible changes yet. Deleters covered by unit tests.
TBR=dbeam@chromium.org

Review URL: https://codereview.chromium.org/124183002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243315 4ff67af0-8c30-449e-8e8b-ad334ec8d88c