mac: Add a missing animation context to bookmark bar controller.
The animation was not specifying an NSAnimationContext or duration, which meant
that the duration was not well defined. I used the same value that I observed
on my machine prior to my change. It seems a bit slow, but presumably it's the
value most users are seeing.

BUG=NONE
TEST=NONE

Review URL: https://codereview.chromium.org/110753004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243383 4ff67af0-8c30-449e-8e8b-ad334ec8d88c