mac: Fix a history swiper bug with magic mouse.
Canceling a history swipe with the magic mouse would fail to hide the swipe
arrow. Previously, the arrow was dismissed in `dealloc`, and I moved the
animation to `dismiss` in a previous change (107343008). Unfortunately, the
logic for magic mouse history swiping was failing to call `dismiss` if the
history swipe was cancelled.

BUG=330312

Review URL: https://codereview.chromium.org/115693012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243417 4ff67af0-8c30-449e-8e8b-ad334ec8d88c