Do not dispatch events to inactive ephemeral apps
With the exception of the chrome.app.runtime.onLaunched() event,
events will only be dispatched to ephemeral apps that are running.
This was implemented by adding a "can_load_ephemeral_apps" flag to
extensions::Event. Only events which enable this flag will have the
ability to load the event page of apps.

BUG=312460
TEST=browser_tests (EphemeralAppBrowserTest.*)

Review URL: https://codereview.chromium.org/118033002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243477 4ff67af0-8c30-449e-8e8b-ad334ec8d88c