Add multi-level undo/redo functions to Bookmark Manager Private API.
New methods to perform an undo and redo of bookmark changes, as well as
obtaining i18n labels and whether there are actually changes to undo or
redo.

undo()
redo()
getUndoInfo()
getRedoInfo()

This commit is only one part of a multi-level undo/redo system for the
bookmark model and there will be an additional commit to take advantage
of these new function in the bmm.

BUG=126092

Review URL: https://codereview.chromium.org/119233004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243613 4ff67af0-8c30-449e-8e8b-ad334ec8d88c