kiosk: Fix StartupAppLauncher::LaunchApp crash.
Speculative fix. This crash happens if we have pending
OnReadyToLaunch job after CleanUp is being called. The only
possible cause seems to be user manually cancels the launch
while the OnReadyToLaunch job is pending. Use timer for the
job and cancels it in CleanUp to fix the crash.

BUG=331105

Review URL: https://codereview.chromium.org/127183002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243726 4ff67af0-8c30-449e-8e8b-ad334ec8d88c