Switch "Downloads" directory path to the per-profile ones for multi-profile support.
All migration code has already landed. This is the final CL that carries out the switch.
By this CL, Files.app in multi-profile session starts showing each profile's Downloads folder,
not the primary profile's one. This is a transient state for the future fully-multiprofile-aware
Files.app (ETA: M34) that shows all profile's folder in a single window.

The following is the summary/background why the path switch is needed:

We had used /home/chronos/user/Download as the path to the "Downloads" folder for all users.
This does not hold for multi-profile session. The "real" path for multiple users are
/home/chronos/u-unique-hash-userA/Downloads
/home/chronos/u-unique-hash-userB/Downloads
...
and /home/chronos/user is just an alias to the user who first signed into the session.

To make all user's folder available simultaneously in one multi-profile session,
we need to stop using the alias, and differentiate all folders by the real path.

BUG=309556

Review URL: https://codereview.chromium.org/48903012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243794 4ff67af0-8c30-449e-8e8b-ad334ec8d88c