Files.app: Make ProgressCenterItemGroup class to manage the states of items.
Previously the logic of progress center is implemented in ProgressCenter.  But
the logic does not follow the item's animation states so another logic is needed
in the ProgressCenterItemPanel and it is getting more complex.

This CL adds ProgressCenterItemGroup to manage the states of items. This class
responsible for creating the summarized item for the close view of progress
center and controling life time error items.

This CL also adds the test of the group. 
It does not start to use the group actually.

BUG=315438
TEST=file_browser_jstest

Review URL: https://codereview.chromium.org/127763002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c