Fixed current OOBE captive portal metrics.
Now metrics are recorded only when a state of a particular network is
changed. If several consecutive checks indicate the same state for a
network, only the first one is recorded. The purpose of this fix is to
remove noize from so-called "lazy detection" from metrics.

BUG=231758
TEST=unit_tests:NetworkPortalDetectorImplTest*

Review URL: https://codereview.chromium.org/99223009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@243945 4ff67af0-8c30-449e-8e8b-ad334ec8d88c