Sync the launch type pref for apps.
This CL syncs the launch type pref for apps by adding a property to
app_specifics.proto. The sync data is processed in ExtensionSyncService.

extensions::SetLaunchType has been changed to take an ExtensionService
so that it is able to sync the pref change when a launch type is updated.

A value has been added to the LaunchType enum to represent an invalid
launch type preference.

BUG=181576
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/93883004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244112 4ff67af0-8c30-449e-8e8b-ad334ec8d88c