Password manager: Enabling Public Suffix List matching on Windows.
The PSL matching seems to work already on Windows, because it only depends on the platform-independent parts of login_database.cc.

After manually verifying that it works (with a Release ToT build), I'd like to enable it on trunk in this CL.

I will make sure that this change does not propagate to beta and stable before it passes all the reviews in the launch bug (linked from the BUG below).

BUG=324291

Review URL: https://codereview.chromium.org/131683003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244172 4ff67af0-8c30-449e-8e8b-ad334ec8d88c