app_shell: Fix PrefsTabHelper dependency of ExtensionHost
The default preference values are fine for app_shell's pages (as they are
for content_shell, which doesn't use PrefsTabHelper).  Delegate out the
creation of PrefsTabHelper so we only create them in Chrome.

With this patch app_shell can launch a renderer before crashing.

BUG=332435
TEST=browser_tests, manual testing that extensions still work

Review URL: https://codereview.chromium.org/132993002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244236 4ff67af0-8c30-449e-8e8b-ad334ec8d88c