Fix apps page footer shrinking/overlap for small window sizes.
This CL shrinks the logo instead of hiding the nav dots and makes the
nav dots shrink to the same size by using the new*** flexbox spec. It
also delays dot layout until the footer logo is loaded as this is
necessary for accurate measurements.

BUG=332350

Review URL: https://codereview.chromium.org/128013002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244486 4ff67af0-8c30-449e-8e8b-ad334ec8d88c