Fix User->image() without ImageSkia representation after login
If the user chose an external avatar image (a downloaded file or a webcam
shot), the avatar image would remain invalid until the external image has
been loaded and decoded. This led to a race on login where the image may
be accessed before the load has completed. The CL fixes this by
substituting a stub image while the load is in progress.

BUG=331365
TEST=Manual

Review URL: https://codereview.chromium.org/136763003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244536 4ff67af0-8c30-449e-8e8b-ad334ec8d88c