Add hook for extension to check sync status.
Adds a new function to preferencesPrivate that returns a list of
unencrypted sync categories. Also changes preferencesPrivate to use a
whitelist instead of being restricted to component extensions.

This is http://crrev.com/118413003 with modified unit tests to not hit
the use-after-free issue seen in the previous patch.

BUG=326678

Review URL: https://codereview.chromium.org/134533005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244608 4ff67af0-8c30-449e-8e8b-ad334ec8d88c