A CL that caused writes to LocalState during startup was causing browser_test crashes because there would often be uncommitted LocalState changes during shutdown.
In that CL a hack was used to fix this - force commits immediately after those changes (which were expected to be one-time only).

But the one-time changes were actually on every startup and potentially high volume so the many commits were causing significant performance regressions.

This CL fixes the root of the problem. Before shutting down the worker threads, we should schedule a commit of LocalState.

BUG=331273,330908,331766,332093,330795

Review URL: https://codereview.chromium.org/134483003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244647 4ff67af0-8c30-449e-8e8b-ad334ec8d88c