[Download] Return DownloadInterruptReason from OnStartedCallback
DownloadUrlParameters::OnStartedCallback is invoked when the server
response is received for a newly started download. If there's an error,
this callback used to receive a net::Error indicating why.

This CL changes the return value to a DownloadInterruptReason so that
the callback can indicate error conditions that affect downloads that
aren't considered errors at the network layer. Also, it avoids having to
call DownloadManager::Start() when the response indicates that a
download is not possible.

Currently downloads that fail to start return a net::Error, but those
that fail after starting return a DownloadInterruptReason. With this
change, all downloads will return a DownloadInterruptReason regardless
of when the download is interrupted.

BUG=7648

Review URL: https://codereview.chromium.org/114193009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244657 4ff67af0-8c30-449e-8e8b-ad334ec8d88c