This is to allow extensions to determine the total length of Javascript blobs.
This cannot AFAIK be done via passing the length in a custom binding, as base::Value does not support int64 a.k.a. 'long long' types.

The total blob length is going to be used in media_galleries_api.cc to construct the SafeMediaMetadataParser in a followup CL. It appears unused in this patch.

BUG=

Review URL: https://codereview.chromium.org/129823003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@244860 4ff67af0-8c30-449e-8e8b-ad334ec8d88c